sources: refactor git source into module.
Extract the Git source into `sources/_git.py`.

LP: #1648906
